+++ sportsdataverse/college_baseball/__init__.py
@@ -1,0 +1,3 @@
+from __future__ import annotations
+
+from sportsdataverse.college_baseball.college_baseball_espn_ext import *  # noqa: F401,F403